Moving cache_refresh to upgrade maintenance action
