--- .github/workflows/j-cache-refresh.yml
@@ -1,33 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-# Copyright 2022 The HuggingFace Authors.
-
-name: jobs/cache_refresh
-on:
-  workflow_dispatch:
-  push:
-    branches:
-      - main
-    paths:
-      - 'jobs/cache_refresh/**'
-      - 'libs/libcommon/**'
-      - '.github/workflows/j-cache-refresh.yml'
-      - '.github/workflows/_quality-python.yml'
-      - '.github/workflows/_unit-tests-python.yml'
-      - 'tools/docker-compose-mongo.yml'
-  pull_request:
-    paths:
-      - 'jobs/cache_refresh/**'
-      - 'libs/libcommon/**'
-      - '.github/workflows/j-cache-refresh.yml'
-      - '.github/workflows/_quality-python.yml'
-      - '.github/workflows/_unit-tests-python.yml'
-      - 'tools/docker-compose-mongo.yml'
-jobs:
-  quality:
-    uses: ./.github/workflows/_quality-python.yml
-    with:
-      working-directory: jobs/cache_refresh
-  unit-tests:
-    uses: ./.github/workflows/_unit-tests-python.yml
-    with:
-      working-directory: jobs/cache_refresh
