+++ chart/templates/_helpers.tpl
@@ -34,10 +34,6 @@ Docker image management
 {{ include "hf.common.images.image" (dict "imageRoot" .Values.images.jobs.mongodbMigration "global" .Values.global.huggingface) }}
 {{- end -}}
 
-{{- define "jobs.cacheRefresh.image" -}}
-{{ include "hf.common.images.image" (dict "imageRoot" .Values.images.jobs.cacheRefresh "global" .Values.global.huggingface) }}
-{{- end -}}
-
 {{- define "jobs.cacheMaintenance.image" -}}
 {{ include "hf.common.images.image" (dict "imageRoot" .Values.images.jobs.cacheMaintenance "global" .Values.global.huggingface) }}
 {{- end -}}
@@ -75,11 +71,6 @@ app.kubernetes.io/component: "{{ include "name" . }}-storage-admin"
 {{- define "labels.mongodbMigration" -}}
 {{ include "hf.labels.commons" . }}
 app.kubernetes.io/component: "{{ include "name" . }}-mongodb-migration"
-{{- end -}}
-
-{{- define "labels.cacheRefresh" -}}
-{{ include "hf.labels.commons" . }}
-app.kubernetes.io/component: "{{ include "name" . }}-cache-refresh"
 {{- end -}}
 
 {{- define "labels.cacheMaintenance" -}}
