+++ jobs/cache_maintenance/src/cache_maintenance/config.py
@@ -5,7 +5,13 @@ from dataclasses import dataclass, field
 from typing import Optional
 
 from environs import Env
-from libcommon.config import CommonConfig, LogConfig, ProcessingGraphConfig, QueueConfig
+from libcommon.config import (
+    CacheConfig,
+    CommonConfig,
+    LogConfig,
+    ProcessingGraphConfig,
+    QueueConfig,
+)
 
 CACHE_MAINTENANCE_ACTION = None
 
@@ -13,6 +19,7 @@ CACHE_MAINTENANCE_ACTION = None
 @dataclass(frozen=True)
 class JobConfig:
     log: LogConfig = field(default_factory=LogConfig)
+    cache: CacheConfig = field(default_factory=CacheConfig)
     queue: QueueConfig = field(default_factory=QueueConfig)
     common: CommonConfig = field(default_factory=CommonConfig)
     graph: ProcessingGraphConfig = field(default_factory=ProcessingGraphConfig)
@@ -24,6 +31,7 @@ class JobConfig:
 
         return cls(
             log=LogConfig.from_env(),
+            cache=CacheConfig.from_env(),
             queue=QueueConfig.from_env(),
             common=CommonConfig.from_env(),
             graph=ProcessingGraphConfig.from_env(),
