--- jobs/cache_refresh/.flake8
@@ -1,5 +1,0 @@
-[flake8]
-# Recommend matching the black line length (119),
-# rather than using the flake8 default of 79:
-max-line-length = 119
-extend-ignore = "E203"
