--- jobs/cache_refresh/.python-version
@@ -1,1 +1,0 @@
-3.9.15
