--- jobs/cache_refresh/pyproject.toml
@@ -1,44 +1,0 @@
-[tool.poetry]
-authors = ["Andrea Soria <andrea@huggingface.co>"]
-description = "Refresh outdated cache entries based on job runner version"
-name = "cache_refresh"
-version = "0.0.1"
-license = "Apache-2.0"
-
-[tool.poetry.dependencies]
-environs = "^9.5.0"
-libcommon = {path = "../../libs/libcommon", develop = true}
-python = "3.9.15"
-
-[tool.poetry.group.dev.dependencies]
-bandit = "^1.7.4"
-black = "^22.12.0"
-flake8 = "^3.9.2"
-isort = "^5.12.0"
-mypy = "^1.0.0"
-pip-audit = "^2.4.14"
-pytest = "^7.2.1"
-pytest-cov = "^2.12.1"
-
-[build-system]
-build-backend = "poetry.core.masonry.api"
-requires = ["poetry-core>=1.0.0"]
-
-[tool.pytest.ini_options]
-filterwarnings = ["ignore::DeprecationWarning"]
-markers = [
-    "wip: tests being developed"
-]
-
-[tool.coverage.run]
-source = ["cache_refresh"]
-
-[tool.isort]
-profile = "black"
-
-[tool.black]
-line-length = 119
-preview = true
-
-[tool.mypy]
-strict = true
